Settings and path split (#58)
* support calls to settings

do nothing, but don't error when calls to firestore settings are made

* ignore first slash in path name

Path names in firestore can be written '/users/aUser' or 'users/aUser'. The preceding slash was resulting in an additional element in the split array which made the length % 2 fail.

* add mock settings

mockSettings in case someone wants to test that their settings are called correctly.